requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.1
altair==5.1.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.3
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.2
attrs==25.3.0
auth0-python==4.9.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.9.0
blis==0.7.10
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.9
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.15.0
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
crewai==0.121.0
crewai-tools==0.45.0
cryptography==45.0.2
curl_cffi==0.11.1
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
docker==7.1.0
docstring_parser==0.16
duckduckgo-search==3.8.5
durationpy==0.10
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.17.1
elasticsearch==8.9.0
email-validator==2.0.0.post2
embedchain==0.1.128
emoji==2.8.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.9
fastavro==1.11.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.42.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.40.2
googleapis-common-protos==1.70.0
gpt4all==1.0.12
gptcache==0.1.44
grpcio==1.63.2
h11==0.16.0
h2==4.1.0
hf-xet==1.1.2
hpack==4.0.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
instructor==1.8.3
iopath==0.1.10
ipykernel==6.25.2
ipython==8.36.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.3.2
jq==1.6.0
json5==0.12.0
json_repair==0.46.0
jsonpatch==1.33
jsonpickle==4.1.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
kubernetes==32.0.1
lancedb==0.22.1
langchain==0.3.25
langchain-cohere==0.3.5
langchain-community==0.3.24
langchain-core==0.3.61
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langcodes==3.3.0
langsmith==0.3.42
layoutparser==0.3.4
litellm==1.68.0
loguru==0.7.2
lxml==4.9.3
Mako==1.3.10
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.101
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.4.4
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.4.2
nltk==3.8.1
nodeenv==1.9.1
numexpr==2.8.5
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.22.0
openai==1.75.0
opencv-python==4.8.0.76
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.1.0
parso==0.8.4
pathspec==0.11.2
pathy==0.10.2
pdf2image==1.16.3
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.18.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone-client==2.2.4
platformdirs==3.10.0
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.26.1
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.8.0
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.1.1
pypdf==5.5.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.401
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.1.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytube==15.0.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==25.1.1
qdrant-client==1.14.2
rapidfuzz==3.3.1
redis==5.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.3.3
schema==0.7.7
scikit-learn==1.3.1
scipy==1.11.3
sentencepiece==0.1.99
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlitedict==2.1.0
srsly==2.4.7
stack-data==0.6.3
starlette==0.45.3
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.9.0
timm==0.9.7
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
toolz==0.12.0
torch==2.7.0
torchvision==0.22.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
unstructured==0.10.16
unstructured.pytesseract==0.3.12
urllib3==2.4.0
uv==0.7.7
uvicorn==0.34.2
uvloop==0.21.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.1.5
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zstandard==0.23.0
